Create the server, web, and test files.
Create the backend server files, the frontend files, and tests for the
backend. Include `package.json` and `.babelrc` configuration files. Also
include an ESLint configuration file.

Use Express.js for frontend hosting, and use Mocha.js for testing.
Firebase is used as the database for the service.Use Handlebars for
templating. The server also exposes documentation files and a public
API.